+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Nothing currently!
 
+## v0.3.0 - 2022-07-31
+
+### Added
+
+* Implemented the `get_archived_threads()` method (by [@cooperwalbrun](https://github.com/cooperwalbrun))
+
+### Changed
+
+* Changed the `stickied` and `closed` fields on the `Thread` model to `is_stickied` and `is_closed` (by [@cooperwalbrun](https://github.com/cooperwalbrun))
+
 ## v0.2.3 - 2022-07-30
 
 ### Added